+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 Both frontend and backend changes and the final version of the application will be documented in this file.
 
+## 0.16.0 - 16.02.2022 [BE]
+
+- Create Add game endpoint
+- Handle completed games and games with new achievements
+
+## 0.15.1 - 15.02.2022 [BE]
+
+- Add global 'api' prefix to requests
+
 ## 0.15.0 - 15.02.2022 [FE]
 
 - Add proxy to backend